Adds powerups to the game.
Player can activate a powerup by pressing 'z' if it is available.
Powerups can be picked up, and a GUI element will indicate when one is available to activate.
Powerup grants the player increased speed and size, as well as invulnerability to damage from invaderers (excluding collision with a boss invaderer).
Note: the player is still vulnerable to losing lives by invaderers passing.