Sprites are now created with a matrix.
The init method of sprites now receive a matrix as input instead of just two choords. This allows for creating the sprite with a specific position and rotation.
Most importantly, it also allows for creation of several sprites with the same matrix, applying changes to said matrix between operations, which facilitates hugely the turtle-graphics behaviour.